+++ Sources/Examples/SwiftUI Examples/LocationOverrideExample.swift
@@ -15,12 +15,16 @@ struct LocationOverrideExample: View {
             Map {
                 /// The location indicator puck position and heading is controlled by the location provider.
                 Puck2D(bearing: .heading)
-            }
-            .onMapTapGesture { context in
-                /// As a demonstration, override location with the last tap coordinate.
-                let direction = provider.location.coordinate.direction(to: context.coordinate)
-                provider.location = Location(coordinate: context.coordinate)
-                provider.heading = Heading(direction: direction, accuracy: 0)
+
+                /// Handle tap on the map.
+                TapInteraction { context in
+                    /// As a demonstration, override location with the last tap coordinate.
+                    let direction = provider.location.coordinate.direction(to: context.coordinate)
+                    provider.location = Location(coordinate: context.coordinate)
+                    provider.heading = Heading(direction: direction, accuracy: 0)
+
+                    return false
+                }
             }
             .onAppear {
                 /// Override the location and Heading provider with Combine publishers.