+++ Sources/Examples/SwiftUI Examples/StandardInteractiveFeaturesExample.swift
@@ -80,9 +80,7 @@ struct StandardInteractiveFeaturesExample: View {
             }
         }
 
-        /// DON'T USE Standard Experimental style in production, it will break over time.
-        /// Currently this feature is in preview.
-        .mapStyle(.standardExperimental(theme: theme, lightPreset: lightPreset, buildingSelectColor: buildingSelectColor))
+        .mapStyle(.standard(theme: theme, lightPreset: lightPreset, colorBuildingSelect: buildingSelectColor))
         .ignoresSafeArea()
         /// Debug panel
         .overlay(alignment: .bottom) {